backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
alembic==1.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9 